backend/app/temp/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
PyJWT
alembic